pyproject.toml
[project]
name = "python-uv-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nitric ==1.2.1",
    "betterproto ==2.0.0b6",
    "huggingface-hub>=0.25.2",
]

[project.optional-dependencies]
ml = [
    "scipy>=1.14.1",
    "torch>=2.4.1",
    "transformers>=4.45.2",
]
llm = [
    "llama-cpp-python>=0.3.1",
]

[tool.uv]
dev-dependencies = [
    "watchdog>=5.0.3",
]
